Enhancement on perspective camera

+++ src/imagesensor/blenderimage.cpp
@@ -87,8 +87,8 @@ void BlenderImage::PostProcess()
 	float* data = (float*)(m_sharedMemory.bytes + m_header_offset + m_header_offset * g_iTileSize * g_iTileSize * 4 * sizeof(float));
 
 	int offset = 0;
-	for (int i = 0; i < m_rendertarget.GetHeight(); ++i)
-		for (int j = 0; j < m_rendertarget.GetWidth(); ++j)
+	for (int i = 0; i < (int)m_rendertarget.GetHeight(); ++i)
+		for (int j = 0; j < (int)m_rendertarget.GetWidth(); ++j)
 		{
 			const Spectrum& c = m_rendertarget.GetColor(j, m_rendertarget.GetHeight() - i - 1);
 			data[offset++] = c.GetR();
